requirements.txt
aiohttp
click
discord.py
python-dotenv